Refactor PDFDocumentOptions
